Move error handler to middleware - add try/catch to notifications service call
